+++ CHANGELOG.md
@@ -29,6 +29,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [platform]: Added AirConditioner device class from Matterbridge.
 - [platform]: Added a Speaker device type.
+- [platform]: Added estimatedEndTime to Rvc animation.
 
 ### Changed